Move utils function into a seperate file and Remove unused files
