--- ImageSegmentationAndPoseEstimation/AlphaPoseLib/HalpeCOCOAPI/PythonAPI/MANIFEST.in
@@ -1,3 +1,0 @@
-recursive-include common *
-include *.pxd
-include *.pyx
