--- ImageSegmentationAndPoseEstimation/AlphaPoseLib/HalpeCOCOAPI/PythonAPI/Makefile
@@ -1,9 +1,0 @@
-all:
-    # install cocohalpetools locally
-	python setup.py build_ext --inplace
-	rm -rf build
-
-install:
-	# install cocohalpetools to the Python site-packages
-	python setup.py build_ext install
-	rm -rf build
